Update
Added logic for checking status of client
