+++ src/server/client.py
@@ -1,5 +1,7 @@
 import socket
 import sys
+import pickle
+import asyncio
 sys.path.append("../")
 import common
 
@@ -9,7 +11,29 @@ class client:
         self._ID: int = ID
         self._Sock = Sock
         self._Address = Addr
+
+    # Protocol will first broadcast a header: size of data. This function pads out the header with ' ' to match expected size.
+
+    def recv(self,Sock: socket) -> tuple:
+        length:int = int(Sock.recv(64))
+        data: str = Sock.recv(length)
+        
+        return (length, data)
     
+    def Process_header(self, Size: int):
+        TARGETS = len(Size) - self._Header
+        DATA = str(self._Header)
+        return DATA.ljust(TARGETS)
+        
+    def Process_metadata(self,data):
+        Serialized_data = pickle.dumps(data)
+        myset:tuple = (len(Serialized_data),data)
+        return myset
+    
+    
+    def Load_metadata(data):
+        return pickle.loads(data)
+
     def __del__(self):
         print(f"Client: -> {self._Address} Disconnected")
         self._Sock.send(common.flags.Disconnect.encode())
